play/pause is implemented (pause icon to be fixed)
