+++ skypro-music/src/components/Player/Player.module.css
@@ -1,0 +1,388 @@
+.bar {
+    position: absolute;
+    bottom: 0;
+    left: 0;
+    width: 100%;
+    background: rgba(28, 28, 28, 0.5);
+}
+
+.content {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: vertical;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: column;
+    flex-direction: column;
+}
+
+.playerProgress {
+    width: 100%;
+    height: 5px;
+    background: #2e2e2e;
+}
+
+.playerBlock {
+    height: 73px;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: horizontal;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: row;
+    flex-direction: row;
+    -webkit-box-pack: justify;
+    -ms-flex-pack: justify;
+    justify-content: space-between;
+    background-color: #1C1C1C;
+}
+
+.player {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: horizontal;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: row;
+    flex-direction: row;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    -webkit-box-pack: start;
+    -ms-flex-pack: start;
+    justify-content: flex-start;
+}
+
+.controls {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: horizontal;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: row;
+    flex-direction: row;
+    padding: 0 27px 0 31px;
+}
+
+.buttonPrev, .buttonPlay, .buttonPause, .buttonNext, .buttonRepeat, .buttonShuffle {
+    padding: 5px;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+}
+
+.buttonPrev {
+    margin-right: 23px;
+
+    &:hover {
+        cursor: pointer;
+    }
+}
+
+.buttonPrevSvg {
+    width: 15px;
+    height: 14px;
+}
+
+.buttonPlay, .buttonPause {
+    margin-right: 23px;
+    cursor: pointer;
+}
+
+.buttonPlaySvg {
+    width: 22px;
+    height: 20px;
+    fill: #d9d9d9;
+}
+
+.buttonPauseSvg {
+    width: 15px;
+    height: 19px;
+    fill: #d9d9d9;
+}
+
+.buttonNext {
+    margin-right: 28px;
+    fill: #a53939;
+
+    &:hover {
+        cursor: pointer;
+    }
+}
+
+.buttonNextSvg {
+    width: 15px;
+    height: 14px;
+    fill: inherit;
+    stroke: #d9d9d9;
+}
+
+.buttonRepeat {
+    margin-right: 24px;
+
+    &:active {
+        fill: #696969;
+        stroke: #ffffff;
+        cursor: pointer;
+    }
+}
+
+.buttonRepeatSvg {
+    width: 18px;
+    height: 12px;
+    fill: transparent;
+    stroke: #696969;
+
+    &:hover {
+        fill: transparent;
+        stroke: #acacac;
+        cursor: pointer;
+    }
+
+    &:active {
+        fill: transparent;
+        stroke: #ffffff;
+        cursor: pointer;
+    }
+}
+
+.buttonShuffle {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+
+    &:active {
+        fill: #696969;
+        stroke: #ffffff;
+        cursor: pointer;
+    }
+}
+
+.buttonShuffleSvg {
+    width: 19px;
+    height: 12px;
+    fill: transparent;
+    stroke: #696969;
+
+    &:hover {
+        fill: transparent;
+        stroke: #acacac;
+        cursor: pointer;
+    }
+
+    &:active {
+        fill: transparent;
+        stroke: #ffffff;
+        cursor: pointer;
+    }
+}
+
+.trackPlay {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: horizontal;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: row;
+    flex-direction: row;
+}
+
+.trackPlayContainer {
+    width: auto;
+    display: -ms-grid;
+    display: grid;
+    -ms-grid-columns: auto 1fr;
+    grid-template-columns: auto 1fr;
+    grid-template-areas: "image author""image album";
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+}
+
+.trackPlayImage {
+    width: 51px;
+    height: 51px;
+    background-color: #313131;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    -webkit-box-pack: center;
+    -ms-flex-pack: center;
+    justify-content: center;
+    margin-right: 12px;
+    grid-row: 1;
+    -ms-grid-row-span: 2;
+    grid-column: 1;
+    grid-area: image;
+}
+
+.trackPlayImageSvg {
+    width: 18px;
+    height: 17px;
+    fill: transparent;
+    stroke: #4e4e4e;
+}
+
+.trackPlayName {
+    grid-row: 1;
+    grid-column: 2;
+    grid-area: author;
+    min-width: 260px;
+}
+
+.trackPlayNameLink {
+    font-style: normal;
+    font-weight: 400;
+    font-size: 16px;
+    line-height: 24px;
+    color: #ffffff;
+    white-space: nowrap;
+}
+
+.trackPlayAuthor {
+    grid-row: 2;
+    grid-column: 2;
+    grid-area: album;
+    min-width: 260px;
+}
+
+.trackPlayAuthorLink {
+    font-style: normal;
+    font-weight: 400;
+    font-size: 13px;
+    line-height: 24px;
+    color: #ffffff;
+}
+
+.trackPlayLikeDis {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: horizontal;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: row;
+    flex-direction: row;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    margin-left: 26%;
+}
+
+.trackPlayLike, .trackPlayDislike {
+    padding: 5px;
+
+    &:active {
+        fill: #696969;
+        stroke: #ffffff;
+        cursor: pointer;
+    }
+}
+
+.trackPlayLikeSvg {
+    width: 14px;
+    height: 12px;
+    fill: transparent;
+    stroke: #696969;
+    /* fill: #696969;
+    stroke: #ffffff;
+    cursor: pointer; */
+
+    &:hover {
+        fill: transparent;
+        stroke: #acacac;
+        cursor: pointer;
+    }
+
+    &:active {
+        fill: transparent;
+        stroke: #ffffff;
+        cursor: pointer;
+    }
+}
+
+.trackPlayDislike {
+    margin-left: 28.5px;
+
+    &:active {
+        fill: #696969;
+        stroke: #ffffff;
+        cursor: pointer;
+    }
+}
+
+.trackPlayDislikeSvg {
+    width: 14.34px;
+    height: 13px;
+    fill: transparent;
+    stroke: #696969;
+
+    &:hover {
+        fill: transparent;
+        stroke: #acacac;
+        cursor: pointer;
+    }
+
+    &:active {
+        fill: transparent;
+        stroke: #ffffff;
+        cursor: pointer;
+    }
+}
+
+.volumeBlock {
+    width: auto;
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    padding: 0 92px 0 0;
+}
+
+.volumeContent {
+    display: -webkit-box;
+    display: -ms-flexbox;
+    display: flex;
+    -webkit-box-orient: horizontal;
+    -webkit-box-direction: normal;
+    -ms-flex-direction: row;
+    flex-direction: row;
+    -webkit-box-align: center;
+    -ms-flex-align: center;
+    align-items: center;
+    -webkit-box-pack: end;
+    -ms-flex-pack: end;
+    justify-content: end;
+}
+
+.volumeImage {
+    width: 13px;
+    height: 18px;
+    margin-right: 17px;
+}
+
+.volumeSvg {
+    width: 13px;
+    height: 18px;
+    fill: transparent;
+}
+
+.volumeProgress {
+    width: 109px;
+    cursor: pointer;
+}
+
+.volumeProgressLine {
+    width: 109px;
+    cursor: pointer;
+}
