+++ skypro-music/src/context/TrackContext.tsx
@@ -6,6 +6,8 @@ import React, { createContext, useState } from "react";
 export type TrackContextType = {
   currentTrack?: Track;
   setCurrentTrack: (track: Track) => void;
+  isPlaying: boolean;
+  setIsPlaying: (isPlaying: boolean) => void;
 };
 
 export const TrackContext = React.createContext<TrackContextType | null>(null);
@@ -13,19 +15,11 @@ export const TrackContext = React.createContext<TrackContextType | null>(null);
 const TrackProvider: React.FC<{ children: React.ReactNode }> = ({
   children,
 }) => {
-  const [currentTrack, setCurrentTrack] = React.useState<Track>({
-    _id: 0,
-    name: "",
-    author: "",
-    album: "",
-    duration_in_seconds: 0,
-    release_date: "",
-    genre: [],
-    track_file: "",
-  });
+  const [currentTrack, setCurrentTrack] = React.useState<Track | undefined>(undefined);
+  const [isPlaying, setIsPlaying] = React.useState(false);
 
   return (
-    <TrackContext.Provider value={{ currentTrack, setCurrentTrack }}>
+    <TrackContext.Provider value={{ currentTrack, setCurrentTrack, isPlaying, setIsPlaying }}>
       {children}
     </TrackContext.Provider>
   );
